# src/app/api/mobile/v1/openapi.yaml
openapi: 3.0.0
info:
  title: EduTime Mobile API
  version: 1.0.0
  description: Mobile API for EduTime - Authentication and Resource Management
  contact:
    name: EduTime Support
    email: support@edutime.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.edutime.com
    description: Production server
security:
  - BearerAuth: []
...
paths:
...
  /api/mobile/v1/auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Logout user and revoke refresh token
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: "Refresh token to revoke (optional)"
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login or refresh endpoints
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          type: object
          description: Response data payload
        error:
          $ref: '#/components/schemas/Error'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - UNAUTHORIZED
            - INVALID_TOKEN
            - TOKEN_EXPIRED
            - INVALID_CREDENTIALS
            - EMAIL_NOT_VERIFIED
            - VALIDATION_ERROR
            - NOT_FOUND
            - ALREADY_EXISTS
            - FORBIDDEN
            - CONFLICT
            - RATE_LIMIT_EXCEEDED
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error context or validation errors
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Response timestamp in ISO 8601 format
        requestId:
          type: string
          format: uuid
          description: Unique request identifier for debugging
        pagination:
          type: object
          properties:
            page:
              type: integer
              description: Current page number
            limit:
              type: integer
              description: Items per page
            total:
              type: integer
              description: Total number of items
            totalPages:
              type: integer
              description: Total number of pages
            hasNext:
              type: boolean
              description: Whether there are more pages
            hasPrev:
              type: boolean
              description: Whether there are previous pages
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'